recover pass, reservas, footer, mas iconos

+++ src/front/js/store/flux.js
@@ -14,7 +14,8 @@ const getState = ({ getStore, getActions, setStore }) => {
 			error: "",
 			info: [JSON.parse(localStorage.getItem("info"))],
 			signup: false,
-			reload: false
+			reload: false,
+			reservaConfirmada: localStorage.getItem("infoReserva")
 		},
 		actions: {
 			getHorarios: async () => {
@@ -103,74 +104,6 @@ const getState = ({ getStore, getActions, setStore }) => {
 					});
 				}
 			},
-			// loginEmpresa: async (email, password) => {
-			// 	const store = getStore();
-			// 	try {
-			// 		var myHeaders = new Headers();
-			// 		myHeaders.append("Content-Type", "application/json");
-
-			// 		var raw = JSON.stringify({
-			// 			email: email,
-			// 			password: password
-			// 		});
-
-			// 		var requestOptions = {
-			// 			method: "POST",
-			// 			headers: myHeaders,
-			// 			body: raw,
-			// 			redirect: "follow"
-			// 		};
-
-			// 		const response = await fetch(process.env.BACKEND_URL + "/api/empresa/login", requestOptions);
-			// 		const responseBody = await response.json();
-			// 		if (responseBody.token) {
-			// 			localStorage.setItem("empresa", responseBody.token);
-			// 			localStorage.setItem("login", true);
-			// 			localStorage.setItem("info", JSON.stringify(responseBody));
-			// 			setStore({ info: [responseBody] });
-			// 			setStore({ empresa: true });
-			// 			setStore({ login: true });
-			// 		}
-			// 	} catch (error) {
-			// 		console.log("error", error);
-
-			// 		setStore({ error: error });
-			// 	}
-			// },
-
-			// loginAdmin: async (email, password) => {
-			// 	const store = getStore();
-			// 	try {
-			// 		var myHeaders = new Headers();
-			// 		myHeaders.append("Content-Type", "application/json");
-
-			// 		var raw = JSON.stringify({
-			// 			email: email,
-			// 			password: password
-			// 		});
-
-			// 		var requestOptions = {
-			// 			method: "POST",
-			// 			headers: myHeaders,
-			// 			body: raw,
-			// 			redirect: "follow"
-			// 		};
-			// 		const response = await fetch(process.env.BACKEND_URL + "/api/admin/login", requestOptions);
-			// 		const responseBody = await response.json();
-			// 		if (responseBody.token) {
-			// 			localStorage.setItem("admin", responseBody.token);
-			// 			localStorage.setItem("login", true);
-			// 			localStorage.setItem("info", JSON.stringify(responseBody));
-			// 			setStore({ info: [responseBody] });
-			// 			setStore({ admin: true });
-			// 			setStore({ login: true });
-			// 		}
-			// 	} catch (error) {
-			// 		console.log("error", error);
-
-			// 		setStore({ error: error });
-			// 	}
-			// },
 			logout: () => {
 				const store = getStore();
 				setStore({ admin: localStorage.removeItem("admin") });
@@ -252,30 +185,44 @@ const getState = ({ getStore, getActions, setStore }) => {
 				}
 			},
 			deleteEmpresa: id => {
-				var myHeaders = new Headers();
-				myHeaders.append("Authorization", "Bearer " + localStorage.getItem("admin"));
-				myHeaders.append("Content-Type", "application/json");
+				Swal.fire({
+					title: "Estas seguro?",
+					text: "No podras revertir este cambio!",
+					icon: "warning",
+					showCancelButton: true,
+					confirmButtonColor: "#3085d6",
+					cancelButtonColor: "#d33",
+					cancelButtonText: "Cancelar",
+					confirmButtonText: "Confirmar!"
+				}).then(result => {
+					if (result.isConfirmed) {
+						var myHeaders = new Headers();
+						myHeaders.append("Authorization", "Bearer " + localStorage.getItem("admin"));
+						myHeaders.append("Content-Type", "application/json");
 
-				var raw = "";
+						var raw = "";
 
-				var requestOptions = {
-					method: "DELETE",
-					headers: myHeaders,
-					body: raw,
-					redirect: "follow"
-				};
+						var requestOptions = {
+							method: "DELETE",
+							headers: myHeaders,
+							body: raw,
+							redirect: "follow"
+						};
 
-				fetch(process.env.BACKEND_URL + "/api/empresa/" + id, requestOptions)
-					.then(response => response.text())
-					.then(result => console.log(result))
-					.catch(error => console.log("error", error));
+						fetch(process.env.BACKEND_URL + "/api/empresa/" + id, requestOptions)
+							.then(response => response.text())
+							.then(result => console.log(result))
+							.catch(error => console.log("error", error));
 
-				const store = getStore();
-				const newList = store.empresas.filter(item => item.id !== id);
-				setStore({ empresas: newList });
-				if (newList.length === 0) {
-					setStore({ empresas: [] });
-				}
+						const store = getStore();
+						const newList = store.empresas.filter(item => item.id !== id);
+						setStore({ empresas: newList });
+						if (newList.length === 0) {
+							setStore({ empresas: [] });
+						}
+						Swal.fire("Borrado!", "La empresa ha sido borrada.", "success");
+					}
+				});
 			},
 			editEmpresa: (id, nombre, email) => {
 				var myHeaders = new Headers();
@@ -514,46 +461,64 @@ const getState = ({ getStore, getActions, setStore }) => {
 			},
 			getReservas: async () => {
 				const store = getStore();
-				if (localStorage.getItem("reservas") == null) {
-					try {
-						const response = await fetch(process.env.BACKEND_URL + "/api/reserva");
-						const responseBody = await response.json();
-						setStore({ reservas: responseBody });
-						localStorage.setItem("reservas", JSON.stringify(store.reservas));
-						console.log(responseBody);
-					} catch (error) {
-						console.log(error);
-					}
-				} else {
-					setStore({ reservas: JSON.parse(localStorage.getItem("reservas")) });
+
+				try {
+					const response = await fetch(process.env.BACKEND_URL + "/api/reserva");
+					const responseBody = await response.json();
+					setStore({ reservas: responseBody });
+
+					console.log(responseBody);
+				} catch (error) {
+					console.log(error);
 				}
 			},
-			addReserva: (id_linea, id_horario, id_usuario, asiento) => {
-				var myHeaders = new Headers();
-				myHeaders.append("Content-Type", "application/json");
+			addReserva: async (id_linea, id_horario, id_usuario, asiento) => {
+				localStorage.removeItem("infoReserva");
+				try {
+					var myHeaders = new Headers();
+					myHeaders.append("Authorization", "Bearer " + localStorage.getItem("user"));
+					myHeaders.append("Content-Type", "application/json");
 
-				var raw = JSON.stringify({
-					id_linea: id_linea,
-					id_horario: id_horario,
-					id_usuario: id_usuario,
-					asiento: asiento
-				});
+					var raw = JSON.stringify({
+						id_linea: id_linea,
+						id_horario: id_horario,
+						id_usuario: id_usuario,
+						asiento: asiento
+					});
 
-				var requestOptions = {
-					method: "POST",
-					headers: myHeaders,
-					body: raw,
-					redirect: "follow"
-				};
+					var requestOptions = {
+						method: "POST",
+						headers: myHeaders,
+						body: raw,
+						redirect: "follow"
+					};
 
-				fetch(process.env.BACKEND_URL + "/api/reserva/registrar", requestOptions)
-					.then(response => response.text())
-					.then(result => console.log(result))
-					.catch(error => console.log("error", error));
+					const response = await fetch(process.env.BACKEND_URL + "/api/reserva/", requestOptions);
+					const responseBody = await response.json();
+					if (response.status == 200) {
+						var codigo = Math.floor(Math.random() * 90000) + 10000;
+						setStore({ reload: true });
+						let reservaConfirmada = [
+							{
+								codigo: codigo,
+								id_linea: id_linea,
+								id_horario: id_horario,
+								id_usuario: id_usuario,
+								asiento: asiento
+							}
+						];
+						localStorage.setItem("infoReserva", JSON.stringify(reservaConfirmada));
+						setStore({
+							reservaConfirmada: [localStorage.getItem("infoReserva")]
+						});
+					}
+				} catch (error) {
+					console.log(error);
+				}
 			},
 			editReserva: (id_linea, id_horario, id_usuario, asiento) => {
 				var myHeaders = new Headers();
-				myHeaders.append("Authorization", "Bearer " + localStorage.getItem("token"));
+				myHeaders.append("Authorization", "Bearer " + localStorage.getItem("user"));
 				myHeaders.append("Content-Type", "application/json");
 
 				var raw = JSON.stringify({
@@ -577,7 +542,7 @@ const getState = ({ getStore, getActions, setStore }) => {
 			},
 			deleteReserva: id => {
 				var myHeaders = new Headers();
-				myHeaders.append("Authorization", "Bearer " + localStorage.getItem("token"));
+				myHeaders.append("Authorization", "Bearer " + localStorage.getItem("user"));
 				myHeaders.append("Content-Type", "application/json");
 
 				var raw = "";
@@ -599,6 +564,71 @@ const getState = ({ getStore, getActions, setStore }) => {
 				setStore({ reservas: newList });
 				if (newList.length === 0) {
 					setStore({ reservas: [] });
+				}
+			},
+			solicitudContraseña: async email => {
+				try {
+					var myHeaders = new Headers();
+					myHeaders.append("Content-Type", "application/json");
+
+					var raw = JSON.stringify({
+						email: email
+					});
+
+					var requestOptions = {
+						method: "POST",
+						headers: myHeaders,
+						body: raw,
+						redirect: "follow"
+					};
+					const response = await fetch(process.env.BACKEND_URL + "/api/recuperar", requestOptions);
+					const responseBody = await response.json();
+
+					if (responseBody.msg == "email enviado") {
+						Swal.fire({
+							position: "center",
+							icon: "success",
+							title: "Revise su casilla de email!",
+							showConfirmButton: false,
+							timer: 1500
+						});
+					}
+				} catch (error) {
+					console.log(error);
+				}
+			},
+			resetPassword: async (token, nueva_contraseña) => {
+				const store = getStore();
+				try {
+					var myHeaders = new Headers();
+					myHeaders.append("Content-Type", "application/json");
+
+					var raw = JSON.stringify({
+						token: token,
+						nueva_contraseña: nueva_contraseña
+					});
+
+					var requestOptions = {
+						method: "PUT",
+						headers: myHeaders,
+						body: raw,
+						redirect: "follow"
+					};
+					const response = await fetch(process.env.BACKEND_URL + "/api/resetcontraseña", requestOptions);
+					const responseBody = await response.json();
+
+					if (responseBody.msg == "contraseña cambiada") {
+						setStore({ reload: true });
+						Swal.fire({
+							position: "center",
+							icon: "success",
+							title: "Contraseña cambiada!",
+							showConfirmButton: false,
+							timer: 1500
+						});
+					}
+				} catch (error) {
+					console.log(error);
 				}
 			}
 		}
